LayoutTests/platform/ios-simulator/fast/dom/Window/window-resize-expected.txt

This test makes sure that we have similar behavior as other browsers when resizing and moving a window outside of the screen. This test will fail in some dimensions when run manually because of inconsistencies of when the window can and cannot go behind the doc.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



window.resizeTo Tests

Testing - resizeTo: Too Big
FAIL window.outerWidth should be 320. Was 0.

Testing - resizeTo: Too Small
FAIL window.outerWidth should be 100. Was 0.

Testing - resizeTo: Full Screen Out of Bounds
PASS window.screenY is screen.availTop
FAIL window.outerHeight should be 460. Was 0.
PASS window.screenX is screen.availLeft
FAIL window.outerWidth should be 320. Was 0.

Testing - resizeTo: Not Full Screen Out of Bounds
FAIL window.outerHeight + window.screenY - screen.availTop should be 460. Was 210.
FAIL window.outerWidth + window.screenX - screen.availLeft should be 320. Was 150.

window.resizeBy Tests

Testing - resizeBy: Too Small
FAIL window.outerWidth should be 100. Was 0.

Testing - resizeBy: Too Big
FAIL window.outerWidth should be 320. Was 0.

window.moveTo Tests

Testing - moveTo: Top Left
PASS window.screenY is screen.availTop
PASS window.screenX is screen.availLeft

Testing - moveTo: Top
PASS window.screenY is screen.availTop
PASS window.screenX is screen.availLeft

Testing - moveTo: Top Right
PASS window.screenY is screen.availTop
FAIL window.outerWidth should be 320. Was 0.

Testing - moveTo: Right
PASS window.screenY is screen.availTop
FAIL window.outerWidth should be 320. Was 0.

Testing - moveTo: Bottom Right
FAIL window.outerHeight should be 460. Was 0.
FAIL window.outerWidth should be 320. Was 0.

Testing - moveTo: Bottom
FAIL window.outerHeight should be 460. Was 0.
PASS window.screenX is screen.availLeft

Testing - moveTo: Bottom Left
FAIL window.outerHeight should be 460. Was 0.
PASS window.screenX is screen.availLeft

Testing - moveTo: Left
PASS window.screenY is screen.availTop
PASS window.screenX is screen.availLeft

window.moveBy Tests

Testing - moveBy: Zero Zero
PASS window.screenX == previousScreenX is true
PASS window.screenY == previousScreenY is true

Testing - moveBy: One Zero
PASS window.screenX == previousScreenX + 1 is true
PASS window.screenY == previousScreenY is true

Testing - moveBy: Zero One
PASS window.screenX == previousScreenX is true
PASS window.screenY == previousScreenY + 1 is true

Testing - moveBy: Top Left
PASS window.screenY is screen.availTop
PASS window.screenX is screen.availLeft

Testing - moveBy: Top
PASS window.screenY is screen.availTop
PASS window.screenX is screen.availLeft

Testing - moveBy: Top Right
PASS window.screenY is screen.availTop
FAIL window.outerWidth should be 320. Was 0.

Testing - moveBy: Right
PASS window.screenY is screen.availTop
FAIL window.outerWidth should be 320. Was 0.

Testing - moveBy: Bottom Right
FAIL window.outerHeight should be 460. Was 0.
FAIL window.outerWidth should be 320. Was 0.

Testing - moveBy: Bottom
FAIL window.outerHeight should be 460. Was 0.

Testing - moveBy: Bottom Left
FAIL window.outerHeight should be 460. Was 0.
PASS window.screenX is screen.availLeft

Testing - moveBy: Left
PASS window.screenX is screen.availLeft

PASS successfullyParsed is true

TEST COMPLETE
